requirements.txt
blessed==1.17.12
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-PyMongo==2.3.0
flyingcircus==0.1.3.2
itsdangerous==1.1.0
pymongo==3.11.2
setuptools-scm==5.0.1
Werkzeug==1.0.1
python-dateutil==2.8.1